python_engine/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
langchain==0.3.14
langchain-community==0.3.14
langchain-google-genai==2.0.8
python-dotenv==1.0.1
faiss-cpu
sentence-transformers==3.3.1
pydantic==2.10.4
protobuf==5.29.3
